+++ Sources/Controllers/ALKConversationViewController+TableView.swift
@@ -1,6 +1,6 @@
 //
 //  ALKConversationViewController+TableView.swift
-//  
+//
 //
 //  Created by Mukesh Thawani on 04/05/17.
 //  Copyright © 2017 Applozic. All rights reserved.
@@ -245,30 +245,46 @@ extension ALKConversationViewController: UITableViewDelegate, UITableViewDataSou
                 return cell
             }
         case .genericCard:
-            let cell: ALKCollectionTableViewCell = tableView.dequeueReusableCell(forIndexPath: indexPath)
-            cell.setLocalizedStringFileName(configuration.localizedStringFileName)
-            cell.register(cell: ALKGenericCardCell.self)
-            cell.update(viewModel: message)
-            return cell
-        case .genericList:
-            
-            let cell: ALKGenericListCell = tableView.dequeueReusableCell(forIndexPath: indexPath)
-            cell.setLocalizedStringFileName(configuration.localizedStringFileName)
-            guard let template = viewModel.genericTemplateFor(message: message) as? ALKGenericListTemplate else { return UITableViewCell() }
-            cell.update(template: template)
-            cell.buttonSelected = {[unowned self] tag, title in
-                print("\(title, tag) button selected in generic card")
-                var infoDict = [String: Any]()
-                infoDict["buttonName"] = title
-                infoDict["buttonIndex"] = tag
-                infoDict["template"] = template
-                infoDict["userId"] = self.viewModel.contactId
-                NotificationCenter.default.post(name: Notification.Name(rawValue: "GenericRichListButtonSelected"), object: infoDict)
+            if message.isMyMessage {
+                let cell: ALKMyGenericCardCell = tableView.dequeueReusableCell(forIndexPath: indexPath)
+                cell.setLocalizedStringFileName(configuration.localizedStringFileName)
+                cell.register(cell: ALKGenericCardCell.self)
+                cell.update(viewModel: message)
+                cell.menuAction = {[weak self] action in
+                    self?.menuItemSelected(action: action, message: message) }
+                return cell
+            } else {
+                let cell: ALKFriendGenericCardCell = tableView.dequeueReusableCell(forIndexPath: indexPath)
+                cell.setLocalizedStringFileName(configuration.localizedStringFileName)
+                cell.register(cell: ALKGenericCardCell.self)
+                cell.update(viewModel: message)
+                cell.menuAction = {[weak self] action in
+                    self?.menuItemSelected(action: action, message: message) }
+                return cell
             }
-            return cell
+        case .genericList:
+            if message.isMyMessage {
+                let cell: ALKMyGenericListCell = tableView.dequeueReusableCell(forIndexPath: indexPath)
+                cell.setLocalizedStringFileName(configuration.localizedStringFileName)
+                guard let template = viewModel.genericTemplateFor(message: message) as? [ALKGenericListTemplate] else { return UITableViewCell() }
+                cell.update(viewModel: message)
+                cell.buttonSelected = {[unowned self] tag, title in
+                    self.postGenericListButtonTapNotification(tag: tag, title: title, template: template)
+                }
+                return cell
+            } else {
+                let cell: ALKFriendGenericListCell = tableView.dequeueReusableCell(forIndexPath: indexPath)
+                cell.setLocalizedStringFileName(configuration.localizedStringFileName)
+                guard let template = viewModel.genericTemplateFor(message: message) as? [ALKGenericListTemplate] else { return UITableViewCell() }
+                cell.update(viewModel: message)
+                cell.buttonSelected = {[unowned self] tag, title in
+                    self.postGenericListButtonTapNotification(tag: tag, title: title, template: template)
+                }
+                return cell
+            }
         case .quickReply:
         if message.isMyMessage {
-                
+
                 let cell: ALKMyMessageQuickReplyCell  = tableView.dequeueReusableCell(forIndexPath: indexPath)
                 cell.setLocalizedStringFileName(configuration.localizedStringFileName)
                 cell.register(cell: ALQuickReplyCollectionViewCell.self)
@@ -278,7 +294,7 @@ extension ALKConversationViewController: UITableViewDelegate, UITableViewDataSou
                 cell.menuAction = {[weak self] action in
                     self?.menuItemSelected(action: action, message: message)}
                 return cell
-                
+
             } else {
                 let cell: ALKFriendMessageQuickReplyCell = tableView.dequeueReusableCell(forIndexPath: indexPath)
                 cell.setLocalizedStringFileName(configuration.localizedStringFileName)
@@ -293,7 +309,7 @@ extension ALKConversationViewController: UITableViewDelegate, UITableViewDataSou
                     self?.menuItemSelected(action: action, message: message) }
                 return cell
             }
-        
+
         }
     }
 
@@ -331,7 +347,7 @@ extension ALKConversationViewController: UITableViewDelegate, UITableViewDataSou
     }
 
     public func tableView(_ tableView: UITableView, viewForHeaderInSection section: Int) -> UIView? {
-        
+
         guard let message = viewModel.messageForRow(indexPath: IndexPath(row: 0, section: section)) else {
             return nil
         }
@@ -351,28 +367,25 @@ extension ALKConversationViewController: UITableViewDelegate, UITableViewDataSou
     }
 
     public func tableView(_ tableView: UITableView, willDisplay cell: UITableViewCell, forRowAt indexPath: IndexPath) {
-        
-        let message = viewModel.messageForRow(indexPath: indexPath)
-        
-        guard message?.metadata != nil else {
+
+        guard let message = viewModel.messageForRow(indexPath: indexPath) else {
             return
-            
+        }
+        guard let metadata = message.metadata else {
+            return
         }
 
-        if(message?.messageType == ALKMessageType.quickReply){
-
-            if(message?.isMyMessage)!{
+        if(message.messageType == ALKMessageType.quickReply){
+            if message.isMyMessage {
                 guard let cell =  cell as? ALKMyMessageQuickReplyCell  else {
                     return
                 }
-                
                 cell.setCollectionViewDataSourceDelegate(dataSourceDelegate: self, indexPath: indexPath)
                 let index = cell.collectionView.tag
                 let value = contentOffsetDictionary[index]
                 let horizontalOffset = CGFloat(value != nil ? value!.floatValue : 0)
                 cell.collectionView.setContentOffset(CGPoint(x: horizontalOffset, y: 0), animated: false)
-
-            }else{
+            } else {
                 guard let cell =  cell as? ALKFriendMessageQuickReplyCell else {
                     return
                 }
@@ -381,20 +394,24 @@ extension ALKConversationViewController: UITableViewDelegate, UITableViewDataSou
                 let value = contentOffsetDictionary[index]
                 let horizontalOffset = CGFloat(value != nil ? value!.floatValue : 0)
                 cell.collectionView.setContentOffset(CGPoint(x: horizontalOffset, y: 0), animated: false)
-
             }
-          
-        }else{
-            guard let cell = cell as? ALKCollectionTableViewCell else {
-                        return
+        } else if message.messageType == .genericCard {
+            if message.isMyMessage {
+                guard let cell =  cell as? ALKMyGenericCardCell else {
+                    return
+                }
+                cell.setCollectionViewDataSourceDelegate(dataSourceDelegate: self, indexPath: indexPath)
+                let index = cell.collectionView.tag
+                cell.collectionView.setContentOffset(CGPoint(x: collectionViewOffsetFromIndex(index), y: 0), animated: false)
+            }else{
+                guard let cell =  cell as? ALKFriendGenericCardCell else {
+                    return
+                }
+                cell.setCollectionViewDataSourceDelegate(dataSourceDelegate: self, indexPath: indexPath)
+                let index = cell.collectionView.tag
+                cell.collectionView.setContentOffset(CGPoint(x: collectionViewOffsetFromIndex(index), y: 0), animated: false)
             }
-             cell.setCollectionViewDataSourceDelegate(dataSourceDelegate: self, indexPath: indexPath)
-             let index = cell.collectionView.tag
-            let value = contentOffsetDictionary[index]
-            let horizontalOffset = CGFloat(value != nil ? value!.floatValue : 0)
-            cell.collectionView.setContentOffset(CGPoint(x: horizontalOffset, y: 0), animated: false)
         }
-        
     }
 
     //MARK: Paging
@@ -421,7 +438,7 @@ extension ALKConversationViewController: UITableViewDelegate, UITableViewDataSou
         let minimumDistanceFromTopToTriggerLoadingMore: CGFloat = 200
         let nearTop = distanceFromTop <= minimumDistanceFromTopToTriggerLoadingMore
         if (!nearTop) {return}
-        
+
         self.viewModel.nextPage()
     }
 
@@ -435,6 +452,7 @@ extension ALKConversationViewController: UITableViewDelegate, UITableViewDataSou
             contentOffsetDictionary[collectionView.tag] = horizontalOffset as AnyObject
         }
     }
+    
 }
 
 extension ALTopicDetail: ALKContextTitleDataType {
@@ -470,79 +488,79 @@ extension ALTopicDetail: ALKContextTitleDataType {
 }
 
 extension ALKConversationViewController: UICollectionViewDataSource,UICollectionViewDelegate, UICollectionViewDelegateFlowLayout{
-    
+
     public func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        
+
         guard let message = viewModel.messageForRow(indexPath: IndexPath(row: 0, section: collectionView.tag)),
             let metadata = message.metadata
             else {
                 return 0
         }
-        
+
         if(message.messageType == ALKMessageType.quickReply){
-            
+
             let payload = metadata["payload"] as! String?
-            
+
             let data = payload?.data
-            
+
             do {
                 if let jsonArray = try JSONSerialization.jsonObject(with: data!, options : .allowFragments) as? [Dictionary<String,Any>]{
-                    
+
                     let filteredCustomReqList = jsonArray;
                     return  filteredCustomReqList.count;
-                    
+
                 }
             } catch let error as NSError {
                 print(error)
             }
-            
+
         }else{
-            
+
             guard let collectionView = collectionView as? ALKIndexedCollectionView,
                 let message = viewModel.messageForRow(indexPath: IndexPath(row: 0, section: collectionView.tag)),
                 let template = viewModel.genericTemplateFor(message: message) as? ALKGenericCardTemplate
-                
+
                 else {return 0}
-            
+
             return template.cards.count
-            
+
         }
         return 0
-        
+
     }
-    
+
     public func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        
+
         guard let collectionView = collectionView as? ALKIndexedCollectionView
             else {
                 return UICollectionViewCell()
         }
-        
+
         let message = viewModel.messageForRow(indexPath: IndexPath(row: 0, section: collectionView.tag))
-    
+
         if(message?.messageType == ALKMessageType.quickReply){
-            
+
             guard let dictionary = viewModel.quickReplyDictionary(message: message, indexRow: indexPath.row) else {
                 return  UICollectionViewCell()
             }
-            
+
             let cell: ALQuickReplyCollectionViewCell = collectionView.dequeueReusableCell(forIndexPath: indexPath)
             cell.update(data: dictionary)
-            
+
             cell.buttonSelected = {[weak self] tag, title in
                 self?.viewModel.send(message: title, isOpenGroup: false)
             }
-            
+
             return cell
-            
-        }else{
-            
+
+        }else {
+
             guard let message = viewModel.messageForRow(indexPath: IndexPath(row: 0, section: collectionView.tag)),
                 let template = viewModel.genericTemplateFor(message: message) as? ALKGenericCardTemplate,
                 template.cards.count > indexPath.row else {
                     return UICollectionViewCell()
             }
-            
+
             let cell: ALKGenericCardCell = collectionView.dequeueReusableCell(forIndexPath: indexPath)
             let card = template.cards[indexPath.row]
             cell.update(card: card)
@@ -558,25 +576,28 @@ extension ALKConversationViewController: UICollectionViewDataSource,UICollection
                 NotificationCenter.default.post(name: Notification.Name(rawValue: "GenericRichCardButtonSelected"), object: infoDict)
             }
             return cell
-            
+
         }
     }
-    
-    public func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
-        
-        let message = viewModel.messageForRow(indexPath: IndexPath(row: 0, section: collectionView.tag))
 
-        if(message?.messageType == ALKMessageType.quickReply){
-            
+    public func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
+
+        guard let message = viewModel.messageForRow(indexPath: IndexPath(row: 0, section: collectionView.tag)) else {
+            return CGSize(width: 0, height: 0)
+        }
+        if(message.messageType == ALKMessageType.quickReply){
             guard let dictionary = viewModel.quickReplyDictionary(message: message, indexRow: indexPath.row) else {
                 return  CGSize(width: self.view.frame.width-50, height: 350)
             }
-            
             return viewModel.getSizeForItemAt(row: indexPath.row, withData: dictionary)
-            
+        } else if message.messageType == .genericCard {
+
+            let width = self.view.frame.width - 100 // - 100 to ensure the card appears in the screen
+            let height = ALKGenericCardCollectionView.rowHeightFor(message: message) - 40 // Extra padding for top and bottom
+            return CGSize(width: width, height: height)
         }
         return CGSize(width: self.view.frame.width-50, height: 350)
-        
+
     }
-    
+
 }