+++ Sources/Controllers/ALKConversationViewController.swift
@@ -522,11 +522,12 @@ open class ALKConversationViewController: ALKBaseViewController, Localizable {
         tableView.register(ALKFriendLocationCell.self)
         tableView.register(ALKMyVideoCell.self)
         tableView.register(ALKFriendVideoCell.self)
-        tableView.register(ALKCollectionTableViewCell.self)
-        tableView.register(ALKGenericListCell.self)
+        tableView.register(ALKMyGenericListCell.self)
+        tableView.register(ALKFriendGenericListCell.self)
         tableView.register(ALKFriendMessageQuickReplyCell.self)
         tableView.register(ALKMyMessageQuickReplyCell.self)
-
+        tableView.register(ALKMyGenericCardCell.self)
+        tableView.register(ALKFriendGenericCardCell.self)
     }
 
 
@@ -836,6 +837,22 @@ open class ALKConversationViewController: ALKBaseViewController, Localizable {
             else {return}
         tableView.scrollToRow(at: indexPath, at: .top, animated: true)
 
+    }
+    
+    func postGenericListButtonTapNotification(tag: Int, title: String, template: [ALKGenericListTemplate]) {
+        print("\(title, tag) button selected in generic list")
+        var infoDict = [String: Any]()
+        infoDict["buttonName"] = title
+        infoDict["buttonIndex"] = tag
+        infoDict["template"] = template
+        infoDict["userId"] = self.viewModel.contactId
+        NotificationCenter.default.post(name: Notification.Name(rawValue: "GenericRichListButtonSelected"), object: infoDict)
+    }
+    
+    func collectionViewOffsetFromIndex(_ index: Int) -> CGFloat {
+        let value = contentOffsetDictionary[index]
+        let horizontalOffset = CGFloat(value != nil ? value!.floatValue : 0)
+        return horizontalOffset
     }
 
     fileprivate func hideMediaOptions() {