+++ Sources/ViewModels/ALKConversationViewModel.swift
@@ -265,10 +265,18 @@ open class ALKConversationViewModel: NSObject, Localizable {
             }
             return height
         case .genericCard:
-            return ALKGenericCardCollectionView.rowHeightFor(message: messageModel)
+            if messageModel.isMyMessage {
+                return ALKMyGenericCardCell.rowHeightFor(message: messageModel)
+            } else {
+                return ALKFriendGenericCardCell.rowHeightFor(message: messageModel)
+            }
         case .genericList:
-            guard let template = genericTemplateFor(message: messageModel) as? ALKGenericListTemplate else {return 0}
-            return ALKGenericListCell.rowHeightFor(template: template)
+            guard let template = genericTemplateFor(message: messageModel) as? [ALKGenericListTemplate] else {return 0}
+            if messageModel.isMyMessage {
+                return ALKMyGenericListCell.rowHeightFor(template: template, viewModel: messageModel)
+            } else {
+                return ALKFriendGenericListCell.rowHeightFor(template: template, viewModel: messageModel)
+            }
         case .quickReply:
             if messageModel.isMyMessage {
                 let heigh = ALKMyMessageQuickReplyCell.rowHeigh(viewModel: messageModel, width: maxWidth)
@@ -1229,7 +1237,8 @@ open class ALKConversationViewModel: NSObject, Localizable {
             let payload = metadata["payload"] as? String
             else { return nil}
         do {
-            let cardTemplate = try JSONDecoder().decode(ALKGenericCardTemplate.self, from: payload.data)
+            let cards = try JSONDecoder().decode([ALKGenericCard].self, from: payload.data)
+            let cardTemplate = ALKGenericCardTemplate(cards: cards)
             richMessages[message.identifier] = cardTemplate
             return cardTemplate
         } catch(let error) {
@@ -1238,13 +1247,13 @@ open class ALKConversationViewModel: NSObject, Localizable {
         }
     }
 
-    private func getGenericListTemplateFor(message: ALKMessageViewModel) -> ALKGenericListTemplate? {
+    private func getGenericListTemplateFor(message: ALKMessageViewModel) -> [ALKGenericListTemplate]? {
         guard
             let metadata = message.metadata,
             let payload = metadata["payload"] as? String
             else { return nil}
         do {
-            let cardTemplate = try JSONDecoder().decode(ALKGenericListTemplate.self, from: payload.data)
+            let cardTemplate = try JSONDecoder().decode([ALKGenericListTemplate].self, from: payload.data)
             richMessages[message.identifier] = cardTemplate
             return cardTemplate
         } catch(let error) {