+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The changelog for [ApplozicSwift](https://github.com/AppLozic/ApplozicSwift). Al
 - [AL-2923] Added configure for hide start new in empty conversation
 - [AL-2923] Added configure for hide back button in conversation list
 - [AL-2923] Added configure for changing  color in navigation title
+- [Al-3093] Update rich message layout to display message on top of templates.
 
 ### Fixes